checkpoint, more refactoring
